Create rajavikrant.md
I would recommend a text bar showing or starting steps procedure in the analyse screen. As a first time user, I was not able to find how to select an image and I was repeatedly clicking on the Analyse button. So during the first time use, there should be animation stating the steps to add the image or a text box with "Click here to add image".